# api.yaml
openapi: 3.0.0
servers:
  - url: http://localhost:8080
info:
  title: CIRCABC Share
  description: This is a API definition for the CIRCABC Share service.
  license:
    name: 'EUPL V1.2'
    url: 'https://joinup.ec.europa.eu/sites/default/files/inline-files/EUPL%20v1_2%20EN(1).txt'
  version: '0.1'
...
paths:
...
  '/file/{fileShortUrl}/fileInfo':
      parameters:
      - name: fileShortUrl
        in: path
        description: The short url of the file
        required: true
        schema:
          type: string
      get:
        description: Used by INTERNAL and EXTERNAL users to get file's metadata from short url
        operationId: getFileInfo
        tags:
        - File
        responses:
          200:
            description: SUCCESS Returns the File's metadata of the newly-created file
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/FileBasics'
          400:
            description: BAD REQUEST the Error Message will be empty
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/Status'
          401:
            description: UNAUTHORIZED the Error message will be empty
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/Status'
          403:
            description: FORBIDDEN the Error message will be NotAuthorized, IllegalFileSize, DateLiesInPast, UserHasInsufficientSpace, EmptyFileName, WrongEmailStructure, WrongNameStructure, MessageTooLong
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/Status'
          500:
            description: INTERNAL SERVER ERROR the Error Message will be empty
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/Status'
  '/file/{fileID}':
    parameters:
    - name: fileID
      in: path
      description: The id of the file
      required: true
      schema:
        type: string
    get:
      description: Used by INTERNAL and EXTERNAL users to download a shared file
      operationId: getFile
      parameters:
        - name: password
          in: query
          description: 'Password of the file'
          required: false
          schema:
            type: string
            format: base64
      tags:
      - File
      responses:
        200:
          description: SUCCESS Returns the file and sends a mail to original uploader to inform him of the download
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        401:
          description: UNAUTHORIZED the Error message will be empty
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        404:
          description: NOT FOUND the Error Message will be empty
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        500:
          description: INTERNAL SERVER ERROR the Error Message will be empty
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
    delete:
      description: Used by INTERNAL users and ADMIN in order to delete a file
      operationId: deleteFile
      parameters:
        - name: reason
          in: query
          description: 'Reason for deletion of the file'
          required: false
          schema:
            type: string
      tags:
      - File
      security:
        - OpenID:
          - email
      responses:
        200:
          description: SUCCESS Deletes the file content and its meta data, also sends an email to original uploader with the reason of deletion if the reason is given
        401:
          description: UNAUTHORIZED the Error message will be empty
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        403:
          description: FORBIDDEN the Error message will be NotAuthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        404:
          description: NOT FOUND the Error Message will be empty
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        500:
          description: INTERNAL SERVER ERROR the Error Message will be empty
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'   
    put:
      description: Used by ADMIN to update file's metadata (expiration date)
      operationId: updateFile
      tags:
      - File
      security:
        - OpenID:
          - email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FileBasics'
      responses:
        200:
          description: SUCCESS Updates the file content and its meta data
        401:
          description: UNAUTHORIZED the Error message will be empty
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        403:
          description: FORBIDDEN the Error message will be NotAuthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        404:
          description: NOT FOUND the Error Message will be empty
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        500:
          description: INTERNAL SERVER ERROR the Error Message will be empty
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status' 
  '/file/{fileID}/status':
    put:
      tags:
        - File
      summary: Update file status
      operationId: updateStatus
      parameters:
        - in: path
          name: fileID
          required: true
          description: ID of the file to be updated
          schema:
            type: string
      requestBody:
        description: Status update payload
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FileStatusUpdate'
      responses:
        '200':
          description: File status updated successfully
        '400':
          description: Invalid request body
        '404':
          description: File not found
  '/file/fileRequest':
    post:
      description: Used by INTERNAL users in order to request the reservation of space for a file
      operationId: postFileFileRequest
      tags:
      - File
      security:
        - openId:
          - email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FileRequest'
      responses:
        200:
          description: SUCCESS Returns the File ID of the newly-created file
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileResult'
        400:
          description: BAD REQUEST the Error Message will be empty
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        401:
          description: UNAUTHORIZED the Error message will be empty
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        403:
          description: FORBIDDEN the Error message will be NotAuthorized, IllegalFileSize, DateLiesInPast, UserHasInsufficientSpace, EmptyFileName, WrongEmailStructure, WrongNameStructure, MessageTooLong
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        500:
          description: INTERNAL SERVER ERROR the Error Message will be empty
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
...
  '/file/{fileID}/fileRequest/sharedWithDownloadNotification':
    parameters:
      - name: fileID
        in: path
        description: The id of the file
        required: true
        schema:
          type: string
      - name: userEmail
        in: query
        description: The email of the user
        required: true
        schema:
          type: string 
      - name: downloadNotification
        in: query
        description: true if download notification has to be sent
        required: true
        schema:
          type: boolean
    post:
      operationId: postFileSharedWithDownloadNotification
      tags:
      - File
      security:
        - OpenID:
          - email
      responses:
        200:
          description: SUCCESS
        401:
          description: UNAUTHORIZED the Error message will be empty
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        403:
          description: FORBIDDEN the Error message will be NotAuthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        404:
          description: NOT FOUND the Error Message will be either FileNotFound or UserNotFound 
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        500:
          description: INTERNAL SERVER ERROR the Error Message will be empty
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
...
components:
  securitySchemes:
    OpenID:   # <--- Arbitrary name for the security scheme. Used to refer to it from elsewhere.
      type: openIdConnect
      openIdConnectUrl: https://localhost:8080/realms/
...
  schemas:
    Status:
      type: object
      properties:
        code:
          type: integer
          description: Status code
        message:
          type: string
          description: Additional message
      required:
        - code
...
    FileBasics:
      type: object
      properties:
        expirationDate:
          description: Expiration date of file
          type: string
          format: date
        hasPassword:
          description: File is password-protected
          type: boolean
        name:
          description: Filename
          type: string
        size:
          description: Size of file (Bytes)
          type: number
          minimum: 0
      required:
      - expirationDate
      - hasPassword
      - name
      - size
...
    FileRequest:
      allOf:
        - $ref: '#/components/schemas/FileBasics'
        - type: object
          properties:
            password:
              description: Password protecting the file
              type: string
              format: base64
            downloadNotification:
              description: Email notification sent after download if set to true
              type: boolean
            sharedWith:
              description: User IDs this file is shared with
              type: array
              items:
                $ref: '#/components/schemas/Recipient'
              minItems: 1
              maxItems: 10
          required:
            - sharedWith
            - fileBytes
...
    FileResult:
      type: object
      properties:
        fileId:
          description: file Id
          type: string
      required:
      - fileId
...
    FileStatusUpdate:
      type: object
      properties:
        status:
          description: New status of the file
          type: string
          enum: [ 'FROZEN', 'AVAILABLE', 'DELETED'] 
      required:
        - status